+++ src/main/java/br/unitins/tp2/jiujitsu/model/converterjpa/RegiaoConverter.java
@@ -5,16 +5,16 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
-public class RegiaoConverter implements AttributeConverter<Regiao, Integer> {
+public class RegiaoConverter implements AttributeConverter<Regiao, Long> {
 
     @Override
-    public Integer convertToDatabaseColumn(Regiao regiao) {
+    public Long convertToDatabaseColumn(Regiao regiao) {
         return regiao == null ? null : regiao.getId();
 
     }
 
     @Override
-    public Regiao convertToEntityAttribute(Integer id) {
+    public Regiao convertToEntityAttribute(Long id) {
         return Regiao.valueOf(id);
     }
